+++ commands/mCrud.js
@@ -2,28 +2,45 @@ var capitalize = require('./functions');
 module.exports = function (vscode, fs, path, pathdir) {
 
     vscode.window.showInputBox({
-        prompt: "name of controller/model",
-        placeHolder: "Enter controller & model name"
-    }).then(function (val) {
-        if (val.length == 0) {
-            vscode.window.showErrorMessage("Controller & model file name required.");
-            return;
+        prompt: "name of folder",
+        placeHolder: "Enter name of create or choose folder leave empty for default"
+    }).then(function (folderName) {
+        if (folderName.length == 0) {
+            vscode.window.showInformationMessage("Controllers & models main folder selected.");
         }
-        var controllerPath = path.join(pathdir + "/application/controllers", capitalize.capitalize(val)) + ".php";
-        var modelPath = path.join(pathdir + "/application/models", capitalize.capitalize(val)) + "_model.php";
+        vscode.window.showInputBox({
+            prompt: "name of controller/model",
+            placeHolder: "Enter controller & model name"
+        }).then(function (val) {
+            if (val.length == 0) {
+                vscode.window.showErrorMessage("Controller & model file name required.");
+                return;
+            }
+            var controllerDir = pathdir + "/application/controllers/" + folderName;
+            var modelDir = pathdir + "/application/models/" + folderName;
 
-        fs.access(controllerPath, function (controllerExists) {
-            if (!controllerExists) {
-                vscode.window.showWarningMessage("Controller already exists!");
-            } else {
-                fs.access(modelPath, function (modelExists) {
-                    if (!modelExists) {
-                        vscode.window.showWarningMessage("Model already exists!");
-                    } else {
+            var controllerPath = path.join(controllerDir, capitalize.capitalize(val)) + ".php";
+            var modelPath = path.join(modelDir, capitalize.capitalize(val)) + "_model.php";
 
-                        fs.open(modelPath, "w+", function (err, fd) {
-                            if (err) throw err;
-                            fs.writeFileSync(fd, `<?php 
+            fs.access(controllerPath, function (controllerExists) {
+                if (!controllerExists) {
+                    vscode.window.showWarningMessage("Controller file name already exists!");
+                } else {
+                    fs.access(modelPath, function (modelExists) {
+                        if (!modelExists) {
+                            vscode.window.showWarningMessage("Model file name already exists!");
+                        } else {
+                            if (!fs.existsSync(modelDir)) {
+                                fs.mkdirSync(modelDir);
+                                vscode.window.showInformationMessage(folderName + " folder created in models.");
+                            }
+                            if (!fs.existsSync(controllerDir)) {
+                                fs.mkdirSync(controllerDir);
+                                vscode.window.showInformationMessage(folderName + " folder created in controllers.");
+                            }
+                            fs.open(modelPath, "w+", function (err, fd) {
+                                if (err) throw err;
+                                fs.writeFileSync(fd, `<?php 
                 
 defined('BASEPATH') OR exit('No direct script access allowed');
                         
@@ -58,11 +75,11 @@ class ` + capitalize.capitalize(val) + `_model extends CI_Model
 /* End of file ` + capitalize.capitalize(val) + `_model.php */
                     
                                         `);
-                            fs.close(fd);
+                                fs.close(fd);
 
-                            fs.open(controllerPath, "w+", function (err, fd) {
-                                if (err) throw err;
-                                fs.writeFileSync(fd, `<?php 
+                                fs.open(controllerPath, "w+", function (err, fd) {
+                                    if (err) throw err;
+                                    fs.writeFileSync(fd, `<?php 
         
 defined('BASEPATH') OR exit('No direct script access allowed');
         
@@ -71,6 +88,7 @@ class ` + capitalize.capitalize(val) + ` extends CI_Controller
     public function __construct()
     {
         parent::__construct();
+        $this->load->model('` + folderName + `/` + capitalize.capitalize(val) + `_model', '` + val + `');
     }
 
     public function create()
@@ -94,23 +112,22 @@ class ` + capitalize.capitalize(val) + ` extends CI_Controller
     }
 }
         
-    /* End of file  ` + val + `.php */
+    /* End of file  ` + capitalize.capitalize(val) + `.php */
         
                             `);
-                                fs.close(fd);
+                                    fs.close(fd);
 
-                                var controllerOpenPath = vscode.Uri.file(controllerPath); //A request file path
-                                vscode.workspace.openTextDocument(controllerOpenPath).then(function (val) {
-                                    vscode.window.showTextDocument(val);
+                                    var controllerOpenPath = vscode.Uri.file(controllerPath); //A request file path
+                                    vscode.workspace.openTextDocument(controllerOpenPath).then(function (val) {
+                                        vscode.window.showTextDocument(val);
+                                    });
                                 });
                             });
-
-                        });
-                    }
-                });
-
-            }
+                        }
+                    });
+                }
+            });
+            vscode.window.showInformationMessage('Controller & Model created successfully!');
         });
-        vscode.window.showInformationMessage('Controller & Model successfully! ');
     });
 }