+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All changes will be updated here.
 
 
+## [0.0.4] - 2021-02-27
+- Simple Auth Added
+- Simple Auth API Added
+- Folder selection and creation facility added
+- CRUD Modified
+- Controllers Modified
+- Models Modified
+
 ## [0.0.3] - 2021-02-25
 - Bug in CRUD fixed.
 - Bug in Helper fixed.